requirements.txt
asgiref==3.9.1
Django==4.2.24
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
